# mobile/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "每日減脂計畫 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/exercises": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "List Exercises",
        "operationId": "list_exercises_exercises_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "body_region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Body Region"
            }
          },
          {
            "name": "equipment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseOut"
                  },
                  "title": "Response List Exercises Exercises Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Add Exercise",
        "operationId": "add_exercise_exercises_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exercises/{exercise_id}/alternatives": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Exercise Alternatives",
        "operationId": "exercise_alternatives_exercises__exercise_id__alternatives_get",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exercise Id"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseAlternativeOut"
                  },
                  "title": "Response Exercise Alternatives Exercises  Exercise Id  Alternatives Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exercises/{exercise_id}": {
      "patch": {
        "tags": [
          "training"
        ],
        "summary": "Edit Exercise",
        "operationId": "edit_exercise_exercises__exercise_id__patch",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exercise Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "training"
        ],
        "summary": "Remove Exercise",
        "operationId": "remove_exercise_exercises__exercise_id__delete",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exercise Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workout-templates": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "List Templates",
        "operationId": "list_templates_workout_templates_get",
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateOut"
                  },
                  "title": "Response List Templates Workout Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Create Template",
        "operationId": "create_template_workout_templates_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workout-templates/{template_id}/copy": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Copy Template",
        "operationId": "copy_template_workout_templates__template_id__copy_post",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workout-templates/{template_id}": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Read Template",
        "operationId": "read_template_workout_templates__template_id__get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "training"
        ],
        "summary": "Replace Template",
        "operationId": "replace_template_workout_templates__template_id__put",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "training"
        ],
        "summary": "Remove Template",
        "operationId": "remove_template_workout_templates__template_id__delete",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workout-schedule": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Read Schedule",
        "operationId": "read_schedule_workout_schedule_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleEntryOut"
                  },
                  "title": "Response Read Schedule Workout Schedule Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "training"
        ],
        "summary": "Set Schedule",
        "operationId": "set_schedule_workout_schedule_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduleEntryIn"
                },
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleEntryOut"
                  },
                  "title": "Response Set Schedule Workout Schedule Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExerciseAlternativeOut": {
        "properties": {
          "exercise": {
            "$ref": "#/components/schemas/ExerciseOut"
          },
          "recommended": {
            "type": "boolean",
            "title": "Recommended"
          },
          "hint": {
            "type": "string",
            "title": "Hint"
          }
        },
        "type": "object",
        "required": [
          "exercise",
          "recommended",
          "hint"
        ],
        "title": "ExerciseAlternativeOut"
      },
      "ExerciseIn": {
        "properties": {
          "category_id": {
            "type": "string",
            "title": "Category Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "body_region": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "lower_body",
                  "upper_body",
                  "core",
                  "full_body",
                  "mobility"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Region"
          },
          "equipment": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "bodyweight",
                  "machine",
                  "barbell",
                  "dumbbell",
                  "kettlebell",
                  "cable",
                  "resistance_band",
                  "treadmill"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "home",
                  "gym",
                  "both"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "name"
        ],
        "title": "ExerciseIn"
      },
      "ExerciseOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "category_id": {
            "type": "string",
            "title": "Category Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "body_region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Region"
          },
          "equipment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "met": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Met"
          },
          "is_builtin": {
            "type": "boolean",
            "title": "Is Builtin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "name",
          "description",
          "body_region",
          "equipment",
          "location",
          "met",
          "is_builtin"
        ],
        "title": "ExerciseOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScheduleEntryIn": {
        "properties": {
          "weekday": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Weekday"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          }
        },
        "type": "object",
        "required": [
          "weekday",
          "template_id"
        ],
        "title": "ScheduleEntryIn"
      },
      "ScheduleEntryOut": {
        "properties": {
          "weekday": {
            "type": "integer",
            "title": "Weekday"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          }
        },
        "type": "object",
        "required": [
          "weekday",
          "template_id"
        ],
        "title": "ScheduleEntryOut"
      },
...
      "TemplateIn": {
        "properties": {
          "category_id": {
            "type": "string",
            "title": "Category Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 240.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TemplateItemIn"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "name"
        ],
        "title": "TemplateIn"
      },
...
      "TemplateOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "category_id": {
            "type": "string",
            "title": "Category Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "is_builtin": {
            "type": "boolean",
            "title": "Is Builtin"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TemplateItemOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "name",
          "location",
          "duration_min",
          "is_builtin",
          "items"
        ],
        "title": "TemplateOut"
      },
...
    }
  }
}